pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "autspaces"
description = "Autistica Turing Citizen Science Platform"
version = "0.1.0"
readme = "README.md"
authors = ["test.com"]


[tool.poetry.dependencies]
python = "3.8.6"

django = "^2.2.16"
django-split-settings = "^1.0"
django-axes = "^5.10"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-feature-policy = "^3.6"
django-stubs-ext = "^0.1"

psycopg2-binary = "^2.8"
gunicorn = "^20.0"
python-decouple = "^3.3"
bcrypt = "^3.2"
structlog = "^20.1"
django-open-humans = "^0.1.6"
vcrpy = "^4.2.1"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
django-querycount = "^0.7"
django-migration-linter = "^2.4"
django-extra-checks = "^0.8"
django-coverage-plugin = "^1.8"
nplusone = "^1.0"

wemake-python-styleguide = "^0.14"
flake8-pytest-style = "^1.3"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"
nitpick = "^0.23"

pytest = "^6.2"
pytest-django = "^4.1"
pytest-cov = "^2.10"
pytest-randomly = "^3.4"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.0"
pytest-timeout = "^1.4"
django-test-migrations = "^1.1"
hypothesis = "^5.43"

mypy = "^0.790"
django-stubs = "^1.7"

sphinx = "^3.4"
sphinx-autodoc-typehints = "^1.11"
tomlkit = "^0.7"
doc8 = "^0.8"

yamllint = "^1.25"
safety = "^1.10"
dotenv-linter = "^0.2"
polint = "^0.4"
dennis = "^0.9"
dump-env = "^1.2"
ipython = "^7.19"